feat: 修改配置使其能成功

+++ Week_07/14-03/src/main/java/com/kaitoshy/demo/repository/UserRepository.java
@@ -5,6 +5,7 @@ import com.kaitoshy.demo.entity.User;
 import java.sql.SQLException;
 import java.util.List;
 
+
 public interface UserRepository{
     /**
      * Create table if not exist.
@@ -14,7 +15,7 @@ public interface UserRepository{
     void createTableIfNotExists() throws SQLException;
     
    
-    Long insert(User user) throws SQLException;
+    int insert(User user) throws SQLException;
     
    
     List<User> selectAll() throws SQLException;
